Finally made CLITest work
