+++ test/Chirp.CLITest/CLITestCollection.cs
@@ -1,0 +1,6 @@
+﻿namespace Chirp.CLITest;
+
+[CollectionDefinition("Chirp.CLI Collection", DisableParallelization = true)]
+public class CLITestCollection : ICollectionFixture<CLIFixture>
+{
+}
