+++ test/Chirp.CLITest/Chirp.CLITest.csproj
@@ -24,6 +24,7 @@
 
     <ItemGroup>
         <ProjectReference Include="..\..\src\Chirp.CLI\Chirp.CLI.csproj" />
+        <ProjectReference Include="..\..\src\Chirp.CSVDBService\Chirp.CSVDBService.csproj" />
     </ItemGroup>
 
 </Project>
