--- test/Chirp.CLITest/CollectionDefinition.cs
@@ -1,8 +1,0 @@
-﻿
-namespace Chirp.CLITest
-{
-    [CollectionDefinition("Non-Parallel Collection", DisableParallelization = true)]
-    public class CollectionDefinition
-    {
-    }
-}
